+++ noodles-bgzf/CHANGELOG.md
@@ -6,6 +6,8 @@
 
   * bgzf/writer: Immediately flush block upon reaching the max block size.
 
+  * bgzf/writer: Track virtual position (`Writer::virtual_position`).
+
 ## 0.9.0 - 2022-02-17
 
 ### Added